requirements.txt
code .
```

### 3️⃣ **Crie os arquivos:**

**Arquivo: `requirements.txt`**
```
astrapy
python-dotenv
```

**Arquivo: `.env`** (suas credenciais - NÃO commitar no GitHub!)
```
ASTRA_TOKEN=AstraCS:hKTtwUrdLGRXLdzXjCWQUguZ:a2fd600a5875e02063d35972dc9c207b606c96006d344744d877d9fe2878c4a0
ASTRA_DB_ENDPOINT=https://5f4bf9ac-9aaa-48e5-87b2-82bf780e5854-us-east1.apps.astra.datastax.com
ASTRA_KEYSPACE=default_keyspace